refactor signup tests
